Change netvm to store "next" PC address in a variable and commit it after the instruction finishes.  This makes for nicer assembler.
Fix off-by-one bug in branch in testvm.c on the toggledf program.

Change popbp to pop the top of the stack to be a new base pointer.

Change old popbp to popto.

Make pop take an argument: # of entries to pop.

Remove 'nop'.  Can now be done nicely with pop 0, 0, 0, 0.