Make netvm APIs nicer.
Remove most of the linking and loading code.  We will use a new model TBD.
 - Will probably have separate structures for linked programs and programs
   that are in progress.
 - Don't want to patch symbols on every machine.
 - Some of this functionality is definitely required, but this is not a
   good implementation: throw it away. 

Update the matched run time.  This is cleaned up substantially and should
be easier to maintain.  It has much fewer dependencies on the linking/loading
subsystem (actually none).  However it places a larger burden on the loader.
The loader must pre-patch the symbols, co-processors, etc to the program before
it gets associated with the runtime.  Also, the configuration can't change
between different programs run in the runtime.