Compile tested adjustments for assignment operations.
Add type checking for lvalues in assignment statements in PML.

Add immediate versions of MIN/MAX in the netvm.

Code generation for assignment and locator references is code complete.

Need more tests for pml code generation.

Also need tests for new netvm operations.

Allow locators that specify offset but not length.

Perform length checking in the code generation based on field in both read
accesses and in write accesses.  The rule is that length of the assignment is
the minimum of:
 - the length of the rvalue 
 - length of space available in the variable (based on offset if any)
 - the length specified in the operation.

TODO: ensure that 0-length operations generate errors.
TODO: ensure that loads/stores to invalid packet fields generate errors.
TODO: add a parameter to the halt instruction to set the vm->error field.