Add pkt_adj/get_off() and fix bitfield insertion.
Apparently I completely forgot to do bitfield insertion in assignments.
This commit has said insertion and even implements it reasonably efficiently.
It also has the implementation of the functions to adjust packet parse offsets
and a test that generates a TCP packet using both.  This test case also
exercises the zero-length packet creation.  That is, packets are created
with zero bytes of actual payload.  One must use pkt_ins* to make space in the
buffer for new headers, etc.

Also touched up commands and renamed sxt64() to signx64().