Split pfind into pflt and pxtr
Update all documentation and change calling conventions.  pflt is
used to just filter all packets that match a pattern an pxtr is used
to extract patterns in a range.  This makes the calling conventions for
both much cleaner.  By default pxtr's will match a single range of
packets.  This supports the typical use case of wanting to grab a
certain subset and exit processing when done.  It ensures that any
ongoing capture will terminate.  However a '-m' option will let it
match multiple ranges of packets.