Fix parsing of IPv6 options to not exit early
IPv6 options parsing had a logic flip-flop that caused it to only work
correctly when the option being searched for was the first in the list.
This commit addresses the logic flip-flop and adds a test case to
ipreasm (the only function using this search function) to validate it
properly finding the fragment header after a hop-by-hop header.