Add pmerge mode to merge packets by timestamp
This commit adds a new mode to pmerge to allow merging packets from
multiple files in timestamp order.  This assumes that packets in the
source files are already sorted by timestamp.  It can be a useful
mode of operation for merging trace files from different interfaces or
parts of the network into a single trace.  The utility skips packets
that lack timestamps or packets whose timestamps step backwards in time
relative to other packets in the same file.