Fix PML cg bug for generating funcs from inlines
Code generation was not walking the bodies of inlines when marking
functions as being used.  I probably did this to avoid trying to
"generate" inlines since they get expanded, well, inline.  But, this
meant that if an inline invoked a regular function, directly or
indirectly, the walk wouldn't mark the function as called and would thus
not generate its code.  This would lead to a such an inline basically
aborting when the VM tried to call an unresolved pointer.