Fix lack of SIOCGSTAMP on Linux 5.2 or later
Linux 5.2 deprecated SIOCGSTAMP and replaced it w/ SIOCGSTAMP_OLD
and SIOCGSTAMP_NEW.  This patch fixes the compile error building on
a system without the original ioctl().  A future fix will be needed
to use the newer API.  But XPKT doesn't require a timestamp tag
with every packet so omitting is not erroneous; just suboptimal.