Fix x2hpkt printing issue w/ protocol gaps
Found during code review.  If there is a gap between protocol headers,
then the utility should print up through the next header with the
protocol given as the enclosing PDU.  The mistake was that it was
printing fields through the end of the enclosing PDU.  The worst that
can happen here is that the header fields get printed out of order.