Refactor PRP into PDU
For a long time I've never liked the "protocol parse" (as a noun)
terminology to represent a PDU in a packet.  So this change refactors
everything from parsers to documentation to use the term PDU instead of
PRP.  I previously made the distinction between a PDU and it's parse
data.  But having parse data without corresponding PDU data makes little
sense.  Furthermore, the rest of the code makes more sense to read.
For example the starting offset of a PDU us pdu_soff(pdu).   The header
in a PDU is pdu_header(pdu).

Here is the complete list of identifier changes:
  prparse -> pdunit
  prpspec -> pduspec
  PRP -> PDU
  prp -> pdu
  <PROTO>_parse -> <PROTO>_pdu
  prp_insert -> pdu_inject
  prp_init_parse -> pdu_init
  prp_init_base_parse -> pdu_init_base
  prp_insert_parse -> pdu_insert
  prp_remove_parse -> pdu_remove
  fld_get_prp* ->  fld_get_pdu*
  pkb->pdu to pkb->pdus
  struct pdunit -> struct pdu
  struct pdunit_ops -> struct pdu_ops

Note that I renamed pdu_insert() to pdu_inject().  This is because
pdu_insert() really makes sense to insert a PDU into a PDU tree.  The
data insertion/removal is now pdu_inject()/pdu_cut().

This changest also adjusts the copyrights for the source .c and .h
files.

Rename "PDU base" to "PDU root" in documentation which results in a few
API changes:
   * pdu_init_base() -> pdu_init_root()
   * pdu_is_base() -> pdu_is_root()
   * pdu_get_base() -> pdu_get_root()